--- lib/admin/home_screen.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-class AdminHomeScreen extends StatelessWidget {
-  AdminHomeScreen({Key key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Text('Admin Home Screen'),
-    );
-  }
-}
