+++ lib/proposal/card.dart
